rocm: add v5.2 for the entire stack (#31591)

+++ var/spack/repos/builtin/packages/rocm-smi-lib/package.py
@@ -17,13 +17,14 @@ class RocmSmiLib(CMakePackage):
 
     homepage = "https://github.com/RadeonOpenCompute/rocm_smi_lib"
     git = "https://github.com/RadeonOpenCompute/rocm_smi_lib.git"
-    url = "https://github.com/RadeonOpenCompute/rocm_smi_lib/archive/rocm-5.1.3.tar.gz"
+    url = "https://github.com/RadeonOpenCompute/rocm_smi_lib/archive/rocm-5.2.0.tar.gz"
     tags = ["rocm"]
 
-    maintainers = ["srekolam", "arjun-raj-kuppala"]
+    maintainers = ["srekolam", "renjithravindrankannath"]
     libraries = ["librocm_smi64"]
 
     version("master", branch="master")
+    version("5.2.0", sha256="7bce567ff4e087598eace2cae72d24c98b2bcc93af917eafa61ec9d1e8ef4477")
     version("5.1.3", sha256="8a19ce60dc9221545aa50e83e88d8c4be9bf7cde2425cefb13710131dc1d7b1b")
     version("5.1.0", sha256="21b31b43015b77a9119cf4c1d4ff3864f9ef1f34e2a52a38f985a3f710dc5f87")
     version("5.0.2", sha256="a169129e4ecd1cca134039dc1bf91e1b3721768781abfae4ae61fad60a633472")
@@ -110,8 +111,9 @@ class RocmSmiLib(CMakePackage):
 
     @run_after("install")
     def post_install(self):
-        shutil.rmtree(self.prefix.lib)
-        install_tree(self.prefix.rocm_smi, self.prefix)
-        shutil.rmtree(self.prefix.rocm_smi)
-        os.remove(join_path(self.prefix.bin, "rsmiBindings.py"))
-        symlink("../bindings/rsmiBindings.py", join_path(self.prefix.bin, "rsmiBindings.py"))
+        if self.spec.satisfies("@:5.1"):
+            shutil.rmtree(self.prefix.lib)
+            install_tree(self.prefix.rocm_smi, self.prefix)
+            shutil.rmtree(self.prefix.rocm_smi)
+            os.remove(join_path(self.prefix.bin, "rsmiBindings.py"))
+            symlink("../bindings/rsmiBindings.py", join_path(self.prefix.bin, "rsmiBindings.py"))
